Add support for XDG Base Directory specification
Add support for the XDG Base Directory specification as defined at
https://specifications.freedesktop.org/basedir-spec/latest/.

The config file located at $XDG_CONFIG_HOME/minicpan/config will be
checked before looking for ~/.minicpanrc.

$XDG_CACHE_HOME/minicpan will be used as the local minicpan directory if
local is not defined on the command line or in the configuration file.

Remove the check for $config{local} and $config{remote} in App.pm. The
software already defaults to http://www.cpan.org/ if a remote is not
specified.